requirements.txt
accelerate==1.3.0
albumentations==2.0.8
bitsandbytes==0.45.4
diffusers==0.32.2
einops==0.8.1
h5py==3.14.0
imgaug==0.4.0
lightning==2.5.0.post0
matplotlib==3.10.7
numpy==2.3.4
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
packaging==25.0
peft==0.17.1
Pillow==12.0.0
pytorch_lightning==2.5.0.post0
pyvista==0.46.1
PyYAML==6.0.2
PyYAML==6.0.3
rembg==2.0.62
scipy==1.16.3
skimage==0.0
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.48.3
xformers==0.0.29.post3
